refactor: Split codes
